Remove custom LayerNorm in original source
PyTorch does not support setting the bias to be optional for the LayerNorm module. The original source fixed this by writing the LayerNorm from scratch to allow bias to be optional. But since Equinox is able to incorporate this, it has been removed.

Rolling back to default eqx.nn.LayerNorm.